Gnome 49, fix Yahoo, separate refresh interval with a minimum of 5m, recompile schemas on schema change

+++ stocks@infinicode.de/components/stocks/menuStockTicker.js
@@ -4,22 +4,12 @@ import Pango from 'gi://Pango'
 import St from 'gi://St'
 
 import { getStockColorStyleClass, isNullOrEmpty, roundOrDefault } from '../../helpers/data.js'
-import { SettingsHandler, STOCKS_PORTFOLIOS, STOCKS_SHOW_OFF_MARKET_TICKER_PRICES, STOCKS_SYMBOL_PAIRS, STOCKS_TICKER_DISPLAY_VARIATION, STOCKS_TICKER_INTERVAL, STOCKS_TICKER_STOCK_AMOUNT, STOCKS_USE_PROVIDER_INSTRUMENT_NAMES } from '../../helpers/settings.js'
+import { SettingsHandler, STOCKS_PORTFOLIOS, STOCKS_SYMBOL_PAIRS, STOCKS_TICKER_INTERVAL, STOCKS_USE_PROVIDER_INSTRUMENT_NAMES } from '../../helpers/settings.js'
 
 import { Translations } from '../../helpers/translations.js'
 import * as FinanceService from '../../services/financeService.js'
 
 import { FINANCE_PROVIDER, MARKET_STATES } from '../../services/meta/generic.js'
-
-const SETTING_KEYS_TO_REFRESH = [
-  STOCKS_SYMBOL_PAIRS,
-  STOCKS_PORTFOLIOS,
-  STOCKS_TICKER_INTERVAL,
-  STOCKS_SHOW_OFF_MARKET_TICKER_PRICES,
-  STOCKS_TICKER_DISPLAY_VARIATION,
-  STOCKS_TICKER_STOCK_AMOUNT,
-  STOCKS_USE_PROVIDER_INSTRUMENT_NAMES
-]
 
 const TICKER_ITEM_VARIATION = {
   COMPACT: 0,
@@ -42,20 +32,29 @@ export const MenuStockTicker = GObject.registerClass({
     this._toggleDisplayTimeout = null
     this._settingsChangedId = null
     this._showLoadingInfoTimeoutId = null
+    this._destroyed = false
+    this._quoteSummariesCache = null
+    this._lastFetchTime = 0
 
     this._settings = new SettingsHandler()
-
     this._sync().catch(e => log(e))
 
     this.connect('destroy', this._onDestroy.bind(this))
     this.connect('button-press-event', this._onPress.bind(this))
 
     this._settingsChangedId = this._settings.connect('changed', (value, key) => {
-      this._registerTimeout(false)
-
-      if (SETTING_KEYS_TO_REFRESH.includes(key)) {
-        this._sync().catch(e => log(e))
+      // Invalidate cache if symbols changed
+      if (key === STOCKS_SYMBOL_PAIRS || key === STOCKS_PORTFOLIOS || key === STOCKS_USE_PROVIDER_INSTRUMENT_NAMES) {
+        this._lastFetchTime = 0
       }
+
+      // Ticker interval - restart timer instead of syncing
+      if (key === STOCKS_TICKER_INTERVAL) {
+        this._registerTimeout(false)
+        return
+      }
+
+      this._sync().catch(e => log(e))
     })
 
     this._registerTimeout(false)
@@ -76,32 +75,42 @@ export const MenuStockTicker = GObject.registerClass({
   async _sync () {
     const tickerEnabledItems = this._getEnabledSymbols()
 
-    const tickerBatch = this._getBatch(tickerEnabledItems, this._visibleStockIndex, this._settings.ticker_stock_amount)
-
-    if (isNullOrEmpty(tickerBatch)) {
+    if (isNullOrEmpty(tickerEnabledItems)) {
       this._showInfoMessage(Translations.EMPTY_TICKER_TEXT)
       return
     }
 
-    this._showLoadingInfoTimeoutId = setTimeout(this._showInfoMessage.bind(this), 500)
+    const now = Date.now()
+    const refreshIntervalMs = (this._settings.refresh_interval || 15) * 60 * 1000
+    const shouldFetch = !this._quoteSummariesCache || (now - this._lastFetchTime) >= refreshIntervalMs
 
-    const [yahooQuoteSummaries, otherQuoteSummaries] = await Promise.all([
-      FinanceService.getQuoteSummaryList({
-        symbolsWithFallbackName: tickerBatch.filter(item => item.provider === FINANCE_PROVIDER.YAHOO).map(symbolData => ({ ...symbolData, fallbackName: symbolData.name })),
-        provider: FINANCE_PROVIDER.YAHOO
-      }),
+    if (shouldFetch) {
+      this._lastFetchTime = now
+      this._showLoadingInfoTimeoutId = setTimeout(this._showInfoMessage.bind(this), 500)
 
-      tickerBatch.filter(item => item.provider !== FINANCE_PROVIDER.YAHOO).map(symbolData => FinanceService.getQuoteSummary({
-        ...symbolData,
-        fallbackName: symbolData.name
-      }))
-    ])
+      const [yahooQuoteSummaries, otherQuoteSummaries] = await Promise.all([
+        FinanceService.getQuoteSummaryList({
+          symbolsWithFallbackName: tickerEnabledItems.filter(item => item.provider === FINANCE_PROVIDER.YAHOO).map(symbolData => ({ ...symbolData, fallbackName: symbolData.name })),
+          provider: FINANCE_PROVIDER.YAHOO
+        }),
 
-    const wildMixOfQuoteSummaries = [...yahooQuoteSummaries, ...otherQuoteSummaries]
+        tickerEnabledItems.filter(item => item.provider !== FINANCE_PROVIDER.YAHOO).map(symbolData => FinanceService.getQuoteSummary({
+          ...symbolData,
+          fallbackName: symbolData.name
+        }))
+      ])
 
-    clearTimeout(this._showLoadingInfoTimeoutId)
+      clearTimeout(this._showLoadingInfoTimeoutId)
 
-    this._createMenuTicker({ tickerBatch, quoteSummaries: wildMixOfQuoteSummaries })
+      if (this._destroyed) {
+        return
+      }
+
+      this._quoteSummariesCache = [...yahooQuoteSummaries, ...otherQuoteSummaries]
+    }
+
+    const tickerBatch = this._getBatch(tickerEnabledItems, this._visibleStockIndex, this._settings.ticker_stock_amount)
+    this._createMenuTicker({ tickerBatch, quoteSummaries: this._quoteSummariesCache })
   }
 
   _createMenuTicker ({ tickerBatch, quoteSummaries }) {
@@ -269,6 +278,10 @@ export const MenuStockTicker = GObject.registerClass({
   }
 
   _showInfoMessage (message) {
+    if (this._destroyed) {
+      return
+    }
+
     this.destroy_all_children()
 
     const infoMessageBin = new St.Bin({
@@ -289,6 +302,7 @@ export const MenuStockTicker = GObject.registerClass({
     const buttonID = event.get_button()
 
     if (buttonID === 2 || buttonID === 3) {
+      this._lastFetchTime = 0 // Force refresh on next sync
       this._registerTimeout()
 
       // avoid propagation
@@ -317,6 +331,8 @@ export const MenuStockTicker = GObject.registerClass({
   }
 
   _onDestroy () {
+    this._destroyed = true
+
     if (this._toggleDisplayTimeout) {
       clearInterval(this._toggleDisplayTimeout)
     }
